requirements.txt
# Core
pyspark==3.5.0
psycopg2-binary==2.9.9

# Airflow core
apache-airflow==2.9.0

# Airflow providers (for Spark & Postgres integrating)
apache-airflow-providers-apache-spark==4.1.0
apache-airflow-providers-postgres==5.10.0

# Utils
pandas==2.2.2
numpy==1.26.4

